refactor: move the selection gadgets into composer::select
Pure code motion: `component_boolean` and the four `component_select*`
entry points move out of `composer.rs` into a `select` submodule.
