+++ src/composer.rs
@@ -23,6 +23,7 @@ mod gate;
 mod logic;
 mod point;
 mod range;
+mod select;
 mod truncate;
 
 #[cfg(test)]
@@ -429,26 +430,6 @@ impl Composer {
         self.append_gate(constraint);
     }
 
-    /// Adds a boolean constraint (also known as binary constraint) where the
-    /// gate eq. will enforce that the [`Witness`] received is either `0` or `1`
-    /// by adding a constraint in the circuit.
-    ///
-    /// Note that using this constraint with whatever [`Witness`] that
-    /// is not representing a value equalling 0 or 1, will always force the
-    /// equation to fail.
-    pub fn component_boolean(&mut self, a: Witness) {
-        let zero = Self::ZERO;
-        let constraint = Constraint::new()
-            .mult(1)
-            .output(-BlsScalar::one())
-            .a(a)
-            .b(a)
-            .c(a)
-            .d(zero);
-
-        self.append_gate(constraint);
-    }
-
     /// Decomposes `scalar` into an array truncated to `N` bits (max 256) in
     /// little endian.
     /// The `scalar` for 4, for example, would be deconstructed into the array
@@ -491,115 +472,6 @@ impl Composer {
         self.assert_equal(acc, scalar);
 
         decomposition
-    }
-
-    /// Conditionally selects a [`Witness`] based on an input bit.
-    ///
-    /// bit == 1 => a,
-    /// bit == 0 => b,
-    ///
-    /// `bit` is expected to be constrained by
-    /// [`Composer::component_boolean`]
-    pub fn component_select(
-        &mut self,
-        bit: Witness,
-        a: Witness,
-        b: Witness,
-    ) -> Witness {
-        // bit * a
-        let constraint = Constraint::new().mult(1).a(bit).b(a);
-        let bit_times_a = self.gate_mul(constraint);
-
-        // 1 - bit
-        let constraint =
-            Constraint::new().left(-BlsScalar::one()).constant(1).a(bit);
-        let one_min_bit = self.gate_add(constraint);
-
-        // (1 - bit) * b
-        let constraint = Constraint::new().mult(1).a(one_min_bit).b(b);
-        let one_min_bit_b = self.gate_mul(constraint);
-
-        // [ (1 - bit) * b ] + [ bit * a ]
-        let constraint = Constraint::new()
-            .left(1)
-            .right(1)
-            .a(one_min_bit_b)
-            .b(bit_times_a);
-        self.gate_add(constraint)
-    }
-
-    /// Conditionally selects a [`Witness`] based on an input bit.
-    ///
-    /// bit == 1 => value,
-    /// bit == 0 => 1,
-    ///
-    /// `bit` is expected to be constrained by
-    /// [`Composer::component_boolean`]
-    pub fn component_select_one(
-        &mut self,
-        bit: Witness,
-        value: Witness,
-    ) -> Witness {
-        let b = self[bit];
-        let v = self[value];
-
-        let f_x = BlsScalar::one() - b + (b * v);
-        let f_x = self.append_witness(f_x);
-
-        let constraint = Constraint::new()
-            .mult(1)
-            .left(-BlsScalar::one())
-            .output(-BlsScalar::one())
-            .constant(1)
-            .a(bit)
-            .b(value)
-            .c(f_x);
-
-        self.append_gate(constraint);
-
-        f_x
-    }
-
-    /// Conditionally selects a [`WitnessPoint`] based on an input bit.
-    ///
-    /// bit == 1 => a,
-    /// bit == 0 => b,
-    ///
-    /// `bit` is expected to be constrained by
-    /// [`Composer::component_boolean`]
-    ///
-    /// This mux deliberately stays untyped: it is not a group operation, and
-    /// selecting between unvalidated points is legitimate. Consequently the
-    /// output is only as validated as the chosen input — where the boundary
-    /// rule on [`TorsionFreeWitnessPoint`] requires membership, establish it
-    /// on the selected point.
-    pub fn component_select_point(
-        &mut self,
-        bit: Witness,
-        a: WitnessPoint,
-        b: WitnessPoint,
-    ) -> WitnessPoint {
-        let x = self.component_select(bit, *a.x(), *b.x());
-        let y = self.component_select(bit, *a.y(), *b.y());
-
-        WitnessPoint::new(x, y)
-    }
-
-    /// Conditionally selects a [`Witness`] based on an input bit.
-    ///
-    /// bit == 1 => value,
-    /// bit == 0 => 0,
-    ///
-    /// `bit` is expected to be constrained by
-    /// [`Composer::component_boolean`]
-    pub fn component_select_zero(
-        &mut self,
-        bit: Witness,
-        value: Witness,
-    ) -> Witness {
-        let constraint = Constraint::new().mult(1).a(bit).b(value);
-
-        self.gate_mul(constraint)
     }
 
     /// Evaluate and return `o` by appending a new constraint into the circuit.
